requirements.txt
argparse==1.2.1
mysql-connector-python==2.0.4
py==1.4.30
pyTelegramBotAPI==0.2.9
pytest==2.7.2
requests==2.7.0
six==1.9.0
wsgiref==0.1.2
